backend_python/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
pydantic
pydantic-settings
passlib[bcrypt]
python-jose[cryptography]
python-multipart
python-dotenv
openai
numpy
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
